third_party/blink/web_tests/external/wpt/webrtc-encoded-transform/script-metadata-transform.https-expected.txt

This is a testharness.js-based test.
[FAIL] video metadata: captureTime
  assert_equals: expected "number" but got "undefined"
Harness: the test ran to completion.
